feat: 친구 목록 조회 무한스크롤 API 수정
