+++ Api/src/main/java/tify/server/api/user/service/RetrieveNeighborListUseCase.java
@@ -8,14 +8,13 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.transaction.annotation.Transactional;
 import tify.server.api.common.slice.SliceResponse;
-import tify.server.api.user.model.dto.response.RetrieveNeighborDTO;
 import tify.server.api.utils.UserUtils;
 import tify.server.core.annotation.UseCase;
 import tify.server.domain.domains.user.adaptor.NeighborAdaptor;
 import tify.server.domain.domains.user.adaptor.UserAdaptor;
 import tify.server.domain.domains.user.domain.Neighbor;
 import tify.server.domain.domains.user.dto.condition.NeighborCondition;
-import tify.server.domain.domains.user.dto.model.NeighborVo;
+import tify.server.domain.domains.user.dto.model.RetrieveNeighborDTO;
 
 @UseCase
 @RequiredArgsConstructor
@@ -34,7 +33,8 @@ public class RetrieveNeighborListUseCase {
                                 .map(Neighbor::getToUserId)
                                 .toList());
         NeighborCondition neighborCondition = new NeighborCondition(neighbors, pageable);
-        Slice<NeighborVo> neighborVos = neighborAdaptor.searchNeighbors(neighborCondition);
-        return SliceResponse.of(neighborVos.map(RetrieveNeighborDTO::of));
+        Slice<RetrieveNeighborDTO> neighborDTOS =
+                neighborAdaptor.searchNeighbors(neighborCondition);
+        return SliceResponse.of(neighborDTOS);
     }
 }
