+++ Domain/src/main/java/tify/server/domain/domains/user/repository/NeighborCustomRepository.java
@@ -3,9 +3,9 @@ package tify.server.domain.domains.user.repository;
 
 import org.springframework.data.domain.Slice;
 import tify.server.domain.domains.user.dto.condition.NeighborCondition;
-import tify.server.domain.domains.user.dto.model.NeighborVo;
+import tify.server.domain.domains.user.dto.model.RetrieveNeighborDTO;
 
 public interface NeighborCustomRepository {
 
-    Slice<NeighborVo> searchToPage(NeighborCondition neighborCondition);
+    Slice<RetrieveNeighborDTO> searchToPage(NeighborCondition neighborCondition);
 }
